refactor(domain): split session.go by concern to fix CodeScene hotspot (#808)
session.go had accreted three unrelated concerns into one 954-line file:
the ~250-line SessionMetrics struct + its turn/tool-blocking logic, a
~230-line NLP heuristic for detecting "is the agent waiting on me" text
cues, and the core SessionState/Launcher lifecycle types — plus a single
137-line MergeMetrics function built from ~20 sequential carry-forward
conditionals (CodeScene's Bumpy Road / Large Method pattern). That's the
likely driver of the 6.79 health score.

Split into metrics.go, waiting_cue.go, and a slimmed session.go, mirroring
the file-per-concern convention this package already uses (rate_limit.go,
task_estimate.go, subagent_estimate.go, grouped.go). Decomposed MergeMetrics
into newMergedMetrics + three named carryForward* helpers, one per field
group, instead of one long function. Test files split to match
(metrics_test.go, waiting_cue_test.go, slimmed session_test.go).

Pure move + extract-method: no behavior change, no new imports beyond
stdlib, same package (all exported symbols keep their name and
visibility) — zero-risk for a foundational, high-churn domain file.
